requirements.txt
instagrapi>=2.1.2
python-dotenv>=1.0.1
rich>=13.7.1
pillow>=10.3.0
